requirement.txt
asn1crypto==1.2.0
beautifulsoup4==4.8.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
cx-Oracle==7.2.2
idna==2.8
lxml==4.4.1
Pillow==6.2.1
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
pytesseract==0.3.0
requests==2.22.0
selenium==3.141.0
six==1.13.0
soupsieve==1.9.3
tesseract==0.1.3
tesserocr==2.4.0
urllib3==1.24.2
win-inet-pton==1.1.0
wincertstore==0.2
